refactor customer portal link
